Created Signup Form and view
User registration is handled through the /signup.html endpoint. Displays a custom UserCreationForm.

signupview() checks if the user data being sent is valid and creates a user in the db. Then it redirects to index.html after logging in the user. This can be verified by the base.html header