Initi Gallery/ArtPiece model
A gallery is essentially a queryset of Artpieces that are related to a user via foreignkey. The artpiece is an image file upload which dumps to static/artpieces in root directory.

added URL endpoint for add_artpieceview().

on GET requests, a blank modelform of artpiece is displayed

on POST requests, the post.DATA and post.FILES are handled to create a new object instance in the DB.  Still need to implement Read/Update/Delete