Unit Tests and cleanup
Artpiece should only be a .jpg, .png or .gif file. I have implemented a class of FileField to enforce these restricted, and altered the ArtPiece model to implement this new class.

Added unit testing for the DashboardView

Added error handling for the Addartpiece view. messages now convey the file size and type requirements if a form is not valid.